update R version that we use for docker
Looking at https://circleci.com/gh/benmarwick/rrtools/60, I see that we have the message "ERROR: dependency ‘styler’ is not available for package ‘usethis’". Perhaps this is because we're getting usethis from github, which is newer than our MRAN pkgs, and the MRAN repo might not have this styler pkg yet. Let's see if the MRAN snapshot with the most current version of R has both usethis and styler...