Added packrat functionality
Changed `DESCRIPTION` to Import `packrat` and changed version number.

Added `Dockerfile-with-packrat` to work with `packrat::restore()`. This required removing private packrat libraries on local machine that are symbolic links pointing toward global packrat cache on local machine for `devtools::install('/{{{repo}}}', dep=TRUE)` to work. This also required installing packrat and restoring the private packrat libraries inside the container. This required adding the container's private packrat library path to `.libPaths()` within the same session as `rmarkdown::render()`. On a minor note, I removed blank lines because `
[WARNING]: Empty continuation line found in:
    RUN . /etc/environment   && sudo apt-get update   && sudo apt-get install libudunits2-dev -y   && R -e "devtools::install('/<REPO>', dep=TRUE)"   && R -e "rmarkdown::render('/<REPO>/analysis/paper/paper.Rmd')"
[WARNING]: Empty continuation lines will become errors in a future release.`

If local machine uses `github_dummy` for `github_info`, the Dockerfile would not work with `<REPO>` but does work with `REPO`. That is, `Dockerfile` and `Dockerfile-with-packrat` fail with `/bin/sh: 1: cannot open REPO: No such file`. I changed `<USERNAME>` to `USERNAME` to match this. The only change is line 31.

Also, the Dockerfile is not guaranteed to have `here` package installed (or `devtools` and `git2r` either), all of which are required to knit `paper.Rmd`, so I explicitly install those three packages in `paper.Rmd`.

To use `Dockerfile-with-packrat`, I edited `use_dockerfile()`. First, I added an argument, `package_manager = TRUE` so that the default is to have reproducible package list and versions but one may opt out. This is `package_manager` to facilitate transition to `rstudio/renv` should that project outperform `packrat` in the future. Second, when obtaining the R version that will be pulled from `rocker/verse:r_version`, it asks `utils::sessionInfo()` if packrat not used but reads `pack.lockfile` if packrat used. Third, if `package_manager = TRUE`, `use_dockerfile()` does not use the `Dockerfile` template. Instead, it uses the `Dockerfile-with-packrat` template, installs packrat. A minor change is changing `message()` to `usethis::ui_todo` at the end of `use_dockerfile()`.

`README.Rmd` is updated with a subsection to `6. rrtools::use_dockerfile()` explaining the rationale and some notes behind package management

Build includes one new note: "Unexported object imported by a ':::' call: ‘packrat:::checkPackified’
    See the note in ?`:::` about the use of this operator."

I've added my makeshift test to `tests/testthat/test_use_dockerfile.R` that installs a package not included in my `rocker/verse:3.5.3` image, and tests that functions from that package and one of its Imports work in `paper.Rmd`.