Pre-check git push and remote for errors
If the project already has an `origin` remote associated with the local repo, it stops and warns user rather than trying to overwrite it.
Rather than ask the user to consent to overwriting CONDUCT, CONTRIBUTING, and README, now we rename those files. This also facilitates automated use of function.
Changed 'yes' to "yes" in 107/115.

Unfortunately, lots of whitespace changes. :(